+++ tests/testthat/test-google_reverse_geocode.R
@@ -18,7 +18,7 @@ test_that("location_type is valid", {
   expect_true(
     google_reverse_geocode(location = c(-37, 144),
                            location_type = c("rooftop", "approximate"),
-                           key = 'abc')$error_message == "The provided API key is invalid."
+                           key = 'abc')$error_message == "The provided API key is invalid. "
     )
 
 })
@@ -37,7 +37,7 @@ test_that("result_type is valid",{
     expect_true(
       google_reverse_geocode(location = c(-37.81659, 144.9841),
                              result_type = c("street_address", "route", "intersection"),
-                             key = 'abc')$error_message == "The provided API key is invalid."
+                             key = 'abc')$error_message == "The provided API key is invalid. "
     )
 
 
@@ -63,7 +63,7 @@ test_that("location type set correctly", {
 test_that("reverse geocode attempted", {
 
   expect_true(google_reverse_geocode(location = c(-37.81659, 144.9841),
-                         key = "abc")$error_message == "The provided API key is invalid.")
+                         key = "abc")$error_message == "The provided API key is invalid. ")
 
   expect_true(is.list(google_reverse_geocode(location = c(-37.81659, 144.9841),
                                      key = "abc")$results))